requirements.txt
pydantic>=2.10.3
pydantic-ai>=0.0.12
pytest>=8.3.4
pytest-asyncio>=0.25.0
python-dotenv>=1.0.1
openai>=1.57.0